Adding folder monitoring: -There's now a checkbox in the scan folder pain of the preferences (it would be good to add a tool tip here, but I didn't see an immediate way to do this. I'm sure there is some property that I just missed) -The new class AutoFolderMonitor is started on startup and after the preference dialog is dismissed. It watches for new files and file name changes to the search directory's. -When a change is detected a 1 second timer is started. If in that 1 sec another change is detected the timer is restarted. This keeps the scan from being initiated multiple times when multiple files are added at once. -When the timer reaches 1 second d delegate is called similar to the command line stuff to do a scan and do all.
I also added the ability to reverse the group order in the scan list. This way the missing stuff doesn't show up on top.

Another option would be to have check boxes next to each file to specify if it should be monitored. But then scan all would have to be modified to only scan the monitored folder.